backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Werkzeug==3.0.3
